refactor: InteractionMiniButton 컴포넌트명을 IconButton으로 수정
